refactoring of mission-related stuff to a separate class MissionManager, WIP
