+++ CMakeLists.txt
@@ -44,6 +44,8 @@ include_directories(
 
 add_library(control_interface SHARED
   src/control_interface.cpp
+  src/mission_manager.cpp
+  src/utils.cpp
   )
 
 target_compile_definitions(control_interface
